% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swipeR.R
\name{swipeRwrapper}
\alias{swipeRwrapper}
\title{List of DOM elements for a carousel}
\usage{
swipeRwrapper(...)
}
\arguments{
\item{...}{HTML elements, one for each slide}
}
\value{
A \code{shiny.tag} object.
}
\description{
Enclose a list of DOM elements in a HTML \code{div} element
  to be passed to the \code{\link{swipeR}} function.
}
