% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swiper.R
\name{swiper}
\alias{swiper}
\title{Create a swiper container}
\usage{
swiper(..., width = "600px", height = "300px", loop = TRUE,
  direction = c("horizontal", "vertical"), speed = 400,
  spaceBetween = 100, mousewheel = FALSE, navigation = FALSE,
  pagination = TRUE)
}
\arguments{
\item{...}{Slot for \link{swiperSlide}.}

\item{width}{swiper width in px or percentage.}

\item{height}{swiper height in px or percentage.}

\item{loop}{Whether to play slide in a loop. TRUE by default.}

\item{direction}{Swiper direction: 'horizontal' or 'vertical'}

\item{speed}{Swiper speed (ms).}

\item{spaceBetween}{Space between slides in px, 100 by default.}

\item{mousewheel}{Whether to enable mouse navigation. FALSE by default.}

\item{navigation}{Whether to display arrow for navigation between slides. FALSE by default.}

\item{pagination}{Whether to display bullet for pagination. TRUE by default.}
}
\description{
A swiper container to include in a \link(fluidPage)
}
\note{
See \url{http://idangero.us/swiper/api/#initialize} for the detailed API description.
}
\examples{
if (interactive()) {
 library(shiny)
 library(swipeR)

 shinyApp(
  ui = fluidPage(
   title = "Hello Shiny!",
   swiper(
     direction = "horizontal",
     mousewheel = TRUE,
     width = "100\%",
     height = "800px",
     swiperSlide(img(src = "https://image.flaticon.com/icons/svg/165/165126.svg")),
     swiperSlide(
       tags$iframe(
         src = "http://www.rinterface.com/shiny/tablerDash/",
         width = "100\%",
         frameBorder = "0",
         style = "position:fixed; height: calc(100\% - 10px);"
       )
     ),
     swiperSlide(
      tags$iframe(
        width = "100\%",
        height = "100\%",
        src = "https://www.youtube.com/embed/u-Ofl67lz0g",
        frameborder = "0",
        allow = "accelerometer; autoplay; encrypted-media; gyroscope; picture-in-picture",
        allowfullscreen = NA
      )
     )
   )
  ),
  server = function(input, output) { }
 )
}
}
